requirements.txt
streamlit
gdown
torch
transformers
sentencepiece
python-docx
